Remove adapters util
