+++ ui/src/utils/error.ts
@@ -1,0 +1,26 @@
+import z from "zod";
+
+export function processZodError<T>(error: z.ZodError<T>, init: string[] = []): string[] {
+  return error.errors.reduce((mainAcc, issue): string[] => {
+    switch (issue.code) {
+      case "invalid_union": {
+        return [
+          ...mainAcc,
+          ...issue.unionErrors.reduce(
+            (innerAcc: string[], current: z.ZodError<T>): string[] => [
+              ...innerAcc,
+              ...processZodError(current, mainAcc),
+            ],
+            []
+          ),
+        ];
+      }
+      default: {
+        const errorMsg = issue.path.length
+          ? `${issue.message} at ${issue.path.join(".")}`
+          : issue.message;
+        return [...mainAcc, errorMsg];
+      }
+    }
+  }, init);
+}
